Add Gateway authenticator
