+++ __tests__/index.test.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 
 jest.mock("axios");
 
-import { Authenticator } from "../src/";
+import { CognitoAuthenticator,ApiGatewayAuthenticator } from "../src/";
 
 const DATE = new Date("2017");
 // @ts-ignore
@@ -17,7 +17,7 @@ describe("private functions", () => {
   let authenticator;
 
   beforeEach(() => {
-    authenticator = new Authenticator({
+    authenticator = new CognitoAuthenticator({
       region: "us-east-1",
       userPoolId: "us-east-1_abcdef123",
       userPoolAppId: "123456789qwertyuiop987abcd",
@@ -100,7 +100,7 @@ describe("private functions", () => {
   });
 
   test("should not return cookie domain", async () => {
-    const authenticatorWithNoCookieDomain = new Authenticator({
+    const authenticatorWithNoCookieDomain = new CognitoAuthenticator({
       region: "us-east-1",
       userPoolId: "us-east-1_abcdef123",
       userPoolAppId: "123456789qwertyuiop987abcd",
@@ -223,7 +223,7 @@ describe("private functions", () => {
   });
 });
 
-describe("createAuthenticator", () => {
+describe("createCognitoAuthenticator", () => {
   let params;
 
   beforeEach(() => {
@@ -238,74 +238,74 @@ describe("createAuthenticator", () => {
   });
 
   test("should create authenticator", () => {
-    expect(typeof new Authenticator(params)).toBe("object");
+    expect(typeof new CognitoAuthenticator(params)).toBe("object");
   });
 
   test("should create authenticator without cookieExpirationDay", () => {
     delete params.cookieExpirationDays;
-    expect(typeof new Authenticator(params)).toBe("object");
+    expect(typeof new CognitoAuthenticator(params)).toBe("object");
   });
 
   test("should create authenticator without disableCookieDomain", () => {
     delete params.disableCookieDomain;
-    expect(typeof new Authenticator(params)).toBe("object");
+    expect(typeof new CognitoAuthenticator(params)).toBe("object");
   });
 
   test("should fail when creating authenticator without params", () => {
     // @ts-ignore
     // ts-ignore is used here to override typescript's type check in the constructor
     // this test is still useful when the library is imported to a js file
-    expect(() => new Authenticator()).toThrow("Expected params");
+    expect(() => new CognitoAuthenticator()).toThrow("Expected params");
   });
 
   test("should fail when creating authenticator without region", () => {
     delete params.region;
-    expect(() => new Authenticator(params)).toThrow("region");
+    expect(() => new CognitoAuthenticator(params)).toThrow("region");
   });
 
   test("should fail when creating authenticator without userPoolId", () => {
     delete params.userPoolId;
-    expect(() => new Authenticator(params)).toThrow("userPoolId");
+    expect(() => new CognitoAuthenticator(params)).toThrow("userPoolId");
   });
 
   test("should fail when creating authenticator without userPoolAppId", () => {
     delete params.userPoolAppId;
-    expect(() => new Authenticator(params)).toThrow("userPoolAppId");
+    expect(() => new CognitoAuthenticator(params)).toThrow("userPoolAppId");
   });
 
   test("should fail when creating authenticator without userPoolDomain", () => {
     delete params.userPoolDomain;
-    expect(() => new Authenticator(params)).toThrow("userPoolDomain");
+    expect(() => new CognitoAuthenticator(params)).toThrow("userPoolDomain");
   });
 
   test("should fail when creating authenticator with invalid region", () => {
     params.region = 123;
-    expect(() => new Authenticator(params)).toThrow("region");
+    expect(() => new CognitoAuthenticator(params)).toThrow("region");
   });
 
   test("should fail when creating authenticator with invalid userPoolId", () => {
     params.userPoolId = 123;
-    expect(() => new Authenticator(params)).toThrow("userPoolId");
+    expect(() => new CognitoAuthenticator(params)).toThrow("userPoolId");
   });
 
   test("should fail when creating authenticator with invalid userPoolAppId", () => {
     params.userPoolAppId = 123;
-    expect(() => new Authenticator(params)).toThrow("userPoolAppId");
+    expect(() => new CognitoAuthenticator(params)).toThrow("userPoolAppId");
   });
 
   test("should fail when creating authenticator with invalid userPoolDomain", () => {
     params.userPoolDomain = 123;
-    expect(() => new Authenticator(params)).toThrow("userPoolDomain");
+    expect(() => new CognitoAuthenticator(params)).toThrow("userPoolDomain");
   });
 
   test("should fail when creating authenticator with invalid cookieExpirationDay", () => {
     params.cookieExpirationDays = "123";
-    expect(() => new Authenticator(params)).toThrow("cookieExpirationDays");
+    expect(() => new CognitoAuthenticator(params)).toThrow("cookieExpirationDays");
   });
 
   test("should fail when creating authenticator with invalid disableCookieDomain", () => {
     params.disableCookieDomain = "123";
-    expect(() => new Authenticator(params)).toThrow("disableCookieDomain");
+    expect(() => new CognitoAuthenticator(params)).toThrow("disableCookieDomain");
   });
 });
 
@@ -313,7 +313,7 @@ describe("handle", () => {
   let authenticator;
 
   beforeEach(() => {
-    authenticator = new Authenticator({
+    authenticator = new CognitoAuthenticator({
       region: "us-east-1",
       userPoolId: "us-east-1_abcdef123",
       userPoolAppId: "123456789qwertyuiop987abcd",
@@ -581,3 +581,21 @@ const getCloudfrontRequest = () => ({
     },
   ],
 });
+
+describe("createApiGatewayAuthenticator", () => {
+  let params;
+
+  beforeEach(() => {
+    params = {
+      region: "us-east-1",
+      userPoolId: "us-east-1_abcdef123",
+      userPoolAppId: "123456789qwertyuiop987abcd",
+      userPoolDomain: "my-cognito-domain.auth.us-east-1.amazoncognito.com",
+    };
+  });
+
+  test("should create authenticator", () => {
+    expect(typeof new ApiGatewayAuthenticator(params)).toBe("object");
+  });
+
+})
